requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
browser-cookie3==0.20.1
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dateparser==1.2.1
distro==1.9.0
docker==7.1.0
fastapi==0.115.12
fonttools==4.58.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
lz4==4.4.4
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.6
openai==1.88.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
regex==2024.11.6
requests==2.32.3
selenium==4.33.0
setuptools==80.9.0
shadowcopy==0.0.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.46.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.4.0
uvicorn==0.34.2
websocket-client==1.8.0
websockets==15.0.1
WMI==1.5.1
wsproto==1.2.0
